pyproject.toml
[project]
name = "mohspitality"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "cryptography>=44.0.2",
    "fastapi-mail>=1.4.2",
    "fastapi[standard]>=0.115.12",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "qrcode[pil]>=8.0",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]",
]
